Making more refactoring. Creating evaluation_utils.py and moving functions there
