+++ examples/nlp/intent_detection_slot_tagging/joint_intent_slot_infer.py
@@ -26,11 +26,12 @@ from nemo import logging
 from nemo.collections.nlp.data.datasets.joint_intent_slot_dataset import JointIntentSlotDataDesc
 from nemo.collections.nlp.nm.data_layers import BertJointIntentSlotDataLayer
 from nemo.collections.nlp.nm.trainables.joint_intent_slot import JointIntentSlotClassifier
-from nemo.collections.nlp.utils.callback_utils import (
+from nemo.collections.nlp.utils.callback_utils import get_classification_report, get_f1_scores
+from nemo.collections.nlp.utils.evaluation_utils import (
     analyze_confusion_matrix,
     errors_per_class,
-    get_classification_report,
-    get_f1_scores,
+    log_misclassified_queries,
+    log_misclassified_slots,
 )
 
 
@@ -40,77 +41,6 @@ def concatenate(lists):
 
 def get_preds(logits):
     return np.argmax(logits, 1)
-
-
-def log_misclassified_queries(intent_labels, intent_preds, queries, intent_dict, limit=50):
-    """
-    Display examples of Intent mistakes.
-    In a format: Query, predicted and labeled intent names.
-    """
-    logging.info(f'*** Misclassified intent queries (limit {limit}) ***')
-    cnt = 0
-    for i in range(len(intent_preds)):
-        if intent_labels[i] != intent_preds[i]:
-            query = queries[i].split('\t')[0]
-            logging.info(
-                f'{query} (predicted: {intent_dict[intent_preds[i]]} - labeled: {intent_dict[intent_labels[i]]})'
-            )
-            cnt = cnt + 1
-            if cnt >= limit:
-                break
-
-
-def log_misclassified_slots(
-    intent_labels, intent_preds, slot_labels, slot_preds, subtokens_mask, queries, intent_dict, slot_dict, limit=50
-):
-    """
-    Display examples of Slot mistakes.
-    In a format: Query, predicted and labeled intent names and list of predicted and labeled slot numbers.
-    also prints dictionary of the slots at the start for easier reading.
-    """
-    logging.info('')
-    logging.info(f'*** Misclassified slots queries (limit {limit}) ***')
-    # print slot dictionary
-    logging.info(f'Slot dictionary:')
-    str = ''
-    for i, slot in enumerate(slot_dict):
-        str += f'{i} - {slot}, '
-        if i % 5 == 4 or i == len(slot_dict) - 1:
-            logging.info(str)
-            str = ''
-
-    logging.info('----------------')
-    cnt = 0
-    for i in range(len(intent_preds)):
-        cur_slot_pred = slot_preds[i][subtokens_mask[i]]
-        cur_slot_label = slot_labels[i][subtokens_mask[i]]
-        if not np.all(cur_slot_pred == cur_slot_label):
-            query = queries[i].split('\t')[0]
-            logging.info(
-                f'{query} (predicted: {intent_dict[intent_preds[i]]} - labeled: {intent_dict[intent_labels[i]]})'
-            )
-            logging.info(f'p: {cur_slot_pred}')
-            logging.info(f'l: {cur_slot_label}')
-            cnt = cnt + 1
-            if cnt >= limit:
-                break
-
-
-def check_problematic_slots(slot_preds_list, slot_dict):
-    """ Check non compliance of B- and I- slots for datasets that use such slot encoding. """
-    cnt = 0
-
-    # for sentence in slot_preds:
-    # slots = sentence.split(" ")
-    sentence = slot_preds_list
-    for i in range(len(sentence)):
-        slot_name = slot_dict[int(sentence[i])]
-        if slot_name.startswith("I-"):
-            prev_slot_name = slot_dict[int(sentence[i - 1])]
-            if slot_name[2:] != prev_slot_name[2:]:
-                print("Problem: " + slot_name + " - " + prev_slot_name)
-                cnt += 1
-    print("Total problematic slots: " + str(cnt))
 
 
 # Parsing arguments
@@ -165,7 +95,6 @@ hidden_states = pretrained_bert_model(
 intent_logits, slot_logits = classifier(hidden_states=hidden_states)
 
 ###########################################################################
-
 # Instantiate an optimizer to perform `infer` action
 evaluated_tensors = nf.infer(
     tensors=[
@@ -178,7 +107,6 @@ evaluated_tensors = nf.infer(
     ],
     checkpoint_dir=args.checkpoint_dir,
 )
-
 
 # --- analyse of the results ---
 intent_logits, slot_logits, loss_mask, subtokens_mask, intent_labels, slot_labels_unmasked = [
@@ -231,24 +159,21 @@ log_misclassified_slots(
 )
 
 # analyze confusion matrices
-print('')
 max_pairs = 20
+logging.info('')
 logging.info(f'*** Most Confused Intents (limit {max_pairs}) ***')
 cm = confusion_matrix(intent_labels, intent_preds)
 analyze_confusion_matrix(cm, intent_dict, max_pairs)
 
-print('')
+logging.info('')
 logging.info(f'*** Intent errors per class (in both directions) ***')
 errors_per_class(cm, intent_dict)
 
-print('')
 max_pair = 20
+logging.info('')
 logging.info(f'*** Most Confused Slots (limit {max_pairs}) ***')
 cm = confusion_matrix(slot_labels, slot_preds, np.arange(len(slot_dict)))
 analyze_confusion_matrix(cm, slot_dict, max_pairs)
-
-# does not work well for large matrices
-# print(f'Intent Confusion matrix:\n{cm}')
 
 # check potentially problematic slots - when I- label comes after different B- label
 # check_problematic_slots(slot_labels, slot_dict)
